Made functions general
I changed the signature of the variables to Array{<:Real,1} etc, instead of
Float64. This is to make the system stable if individual routines are called.
The normal running of the code makes use of read_stack which creates a Float64
stack but if you bypass this routine, the combination routines should still
work if the input is Int for instance.